// client-app/src/utils/enums/TextsEnums.ts
export enum TitleEnums {
  TITLE = "Prueba Técnica",
  RESULT = "Resultados",
}

export enum SubTitleEnums {
  SUB_TITLE = "Buscador de videojuegos",
  SEARCH_TITLE = "Buscar por título",
  SCORE = "Puntuación",
}

export enum AdviceEnums {
  LOADING = "Cargando...",
  ERROR = "Error",
  NOT_RESULTS = "No hay resultados",
}

export enum ButtonEnums {
  SEARCH = "Buscar",
}